Leader: Restrict the timer to 8 bits
The default timer fits well into 8 bits, there is no reason at this time
to have more than that for this purpose. Having them at half size
decreases our memory use by a few bytes.

+++ lib/Akela-Leader/src/Akela/Leader.cpp
@@ -24,8 +24,8 @@ namespace Akela {
   // --- state ---
   Key Leader::sequence[LEADER_MAX_SEQUENCE_LENGTH + 1];
   uint8_t Leader::sequencePos;
-  uint16_t Leader::timer;
-  uint16_t Leader::timeOut = DEFAULT_TIMEOUT;
+  uint8_t Leader::timer;
+  uint8_t Leader::timeOut = DEFAULT_TIMEOUT;
   const Leader::dictionary_t *Leader::dictionary;
 
   // --- helpers ---
