+++ lib/Akela-Leader/src/Akela/Leader.h
@@ -48,8 +48,8 @@ namespace Akela {
   private:
     static Key sequence[LEADER_MAX_SEQUENCE_LENGTH + 1];
     static uint8_t sequencePos;
-    static uint16_t timer;
-    static uint16_t timeOut;
+    static uint8_t timer;
+    static uint8_t timeOut;
     static const Leader::dictionary_t *dictionary;
 
     static Key eventHandlerHook (Key mappedKey, byte row, byte col, uint8_t keyState);
